+++ framework/src/test/java/org/checkerframework/framework/test/junit/H1H2CheckerTest.java
@@ -7,7 +7,7 @@ import org.junit.runners.Parameterized.Parameters;
 import java.io.File;
 import java.util.List;
 
-/** */
+/** TODO */
 public class H1H2CheckerTest extends CheckerFrameworkPerDirectoryTest {
 
     /** @param testFiles the files containing test code, which will be type-checked */