+++ docs/CHANGELOG.md
@@ -7,7 +7,7 @@ Version 3.23.0-eisop2 (July 22, 2022)
 
 **Closed issues:**
 
-eisop#308.
+eisop#308, eisop#282.
 
 
 Version 3.23.0-eisop1 (July 14, 2022)